[Kernel] Split achievements backends from manager and interface
